Add class and Change variable name
