+++ Tavisca.Bootcamp.LanguageBasics.Exercise2/Program.cs
@@ -1,9 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 
-namespace Tavisca.Bootcamp.LanguageBasics.Exercise1
+namespace Tavisca.Bootcamp.LanguageBasics.Exercise2
 {
-    public static class Program
+    class Program
     {
         static void Main(string[] args)
         {
@@ -16,49 +15,10 @@ namespace Tavisca.Bootcamp.LanguageBasics.Exercise1
 
         private static void Test(string[] postTimes, string[] showTimes, string expected)
         {
-            var result = GetCurrentTime(postTimes, showTimes).Equals(expected) ? "PASS" : "FAIL";
+            var result = ForumPostEasy.GetCurrentTime(postTimes, showTimes).Equals(expected) ? "PASS" : "FAIL";
             var postTimesCsv = string.Join(", ", postTimes);
             var showTimesCsv = string.Join(", ", showTimes);
             Console.WriteLine($"[{postTimesCsv}], [{showTimesCsv}] => {result}");
-        }
-
-        public static string GetCurrentTime(string[] exactPostTime, string[] showPostTime)
-        {
-            var currTime = new List<TimeSpan>();
-            for (int i=0;i<exactPostTime.Length;i++)
-            {
-                for(int j=i+1;j<exactPostTime.Length;j++)
-                {
-                    if(exactPostTime[i]==exactPostTime[j])
-                    {
-                        if (showPostTime[i] != showPostTime[j])
-                            return "impossible";
-                    }
-                }
-                TimeSpan timeSpan = TimeSpan.Parse(exactPostTime[i]);
-                if (showPostTime[i].Contains("few"))
-                {
-                    TimeSpan time = new TimeSpan(timeSpan.Hours, timeSpan.Minutes, timeSpan.Seconds);
-                    currTime.Add(time);
-                }
-                else if(showPostTime[i].Contains("minutes"))
-                {
-                    string minute = showPostTime[i].Split(' ')[0];
-                    timeSpan = timeSpan.Add(TimeSpan.FromMinutes(double.Parse(minute)));
-                    TimeSpan time = new TimeSpan(timeSpan.Hours, timeSpan.Minutes, timeSpan.Seconds);
-                    currTime.Add(time);
-                }
-                else
-                {
-                    string hours = showPostTime[i].Split(' ')[0];
-                    timeSpan = timeSpan.Add(TimeSpan.FromHours(double.Parse(hours)));
-                    TimeSpan time = new TimeSpan(timeSpan.Hours, timeSpan.Minutes, timeSpan.Seconds);
-                    currTime.Add(time);
-                }
-                
-            }
-            currTime.Sort();
-            return currTime[currTime.Count - 1].ToString();
         }
     }
 }
